Backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.16,<4
python-dotenv>=1.0,<2
groq>=0.30,<1
psycopg2-binary>=2.9,<3
pypdf>=5,<6
Pillow>=10,<12
pytesseract>=0.3.10,<1
